requirements.txt
Pillow>=10.0.0
PySide6>=6.6.0
pytest>=8.0.0


